requirements.txt
# Core
pydantic>=2.0
pyyaml>=6.0
httpx>=0.27.0

# Apify
apify-client>=1.7.0

# Embedding & Clustering
numpy>=1.26.0
hdbscan>=0.8.33
scikit-learn>=1.4.0

# LLM Providers
anthropic>=0.39.0
openai>=1.50.0          # DashScope compatible & DeepSeek

# DingTalk
requests>=2.32.0

# Data
pandas>=2.2.0
openpyxl>=3.1.0

# Utils
tenacity>=8.2.0
python-dotenv>=1.0.0
